# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sorolens API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Sorolens indexes and surfaces data from Soroban smart contracts: events,
    invocations, storage, and per-contract analytics. This spec is the single
    source of truth for the auto-generated Go client (packages/go-client) and
    matches the live routes in apps/api/internal/router/router.go.

    Two namespaces are documented: the original `/api/v1` surface and the
    consistent `/api/v2` surface introduced in issue #144. v2 is additive; v1 is
    unchanged. See docs/api-v2.md for the field-by-field v1 -> v2 mapping.

    Authentication: most read endpoints accept an optional API key via
    `Authorization: Bearer <token>` or `X-API-Key: <token>`. API-key management
    additionally requires an admin identity via role headers
    (`X-User-ID`, `X-GitHub-Login`, `X-User-Sub`, or `X-GitHub-ID`).
  contact:
    name: Sorolens
servers:
  - url: https://api.sorolens.xyz
    description: Production
...
paths:
...
  /api/v1/contracts/{id}:
    get:
      tags: [Contracts]
      summary: Get one contract
      operationId: getContract
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The contract
          headers:
            X-Cache:
              $ref: "#/components/headers/XCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/contracts/{id}/notes:
    get:
      tags: [Contracts]
      summary: List a contract's notes
      description: |
        Returns a contract's markdown notes, newest first. Notes are open to
        read like the rest of the v0.1 surface; a contract with no notes gets
        an empty list rather than an error.
      operationId: listContractNotes
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The contract's notes, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractNoteList"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Contracts]
      summary: Add a note to a contract
      description: |
        Appends a markdown note to a contract. The stored author is the caller
        identity (X-User-ID), which is what later authorizes editing and
        deleting the note. Authoring requires the contributor role.
      operationId: createContractNote
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractNoteRequest"
      responses:
        "201":
          description: The stored note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractNote"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/notes/{noteID}:
    patch:
      tags: [Contracts]
      summary: Rewrite a note
      description: |
        Replaces a note's body. Only the note's author may rewrite it; any
        other contributor gets 403. The stored author is never reassigned.
      operationId: updateContractNote
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/NoteID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractNoteRequest"
      responses:
        "200":
          description: The updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractNote"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Contracts]
      summary: Delete a note
      description: |
        Removes a note. Only the note's author may delete it. Deleting a note
        that is already gone is a no-op, so a retried delete succeeds.
      operationId: deleteContractNote
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/NoteID"
      responses:
        "204":
          description: Note deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Scoped API key. Send as `Authorization: Bearer <token>` (or `X-API-Key`)."
    RoleAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        Role-based auth. Resolve a user identity from one of
        X-User-ID, X-GitHub-Login, X-User-Sub, or X-GitHub-ID and enforce the
        minimum role declared next to this security scheme.
    IdentityHeaderAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Watchlist and contract-group endpoints read the user ID from X-User-ID.
  parameters:
    ContractID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Contract ID (starts with C).
...
    NoteID:
      name: noteID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Contract note ID.
...
  headers:
    XCache:
      description: |
        Response cache status (`HIT` or `MISS`). Present when the API runs
        with Redis caching enabled; cached bodies live for API_CACHE_TTL
        (default 30s) and are purged by writes to the same resource.
      schema:
        type: string
        enum: [HIT, MISS]
  responses:
    Unauthorized:
      description: Missing or invalid credential or identity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ForbiddenRole:
      description: Identity present but role below the required minimum
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RoleError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidInput:
      description: Invalid query or body input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnsupportedMediaType:
      description: Content-Type is not application/json
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    ContractNote:
      type: object
      required: [id, contract_id, author, body, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        contract_id:
          type: string
        author:
          type: string
          description: |
            Identity that wrote the note (X-User-ID). Only that identity may
            rewrite or delete it.
        body:
          type: string
          maxLength: 8192
          description: |
            Markdown source. Raw HTML is escaped rather than interpreted when
            the note is rendered.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ContractNoteRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 8192
          description: Note markdown. A whitespace-only body is rejected.
    ContractNoteList:
      type: object
      required: [contract_id, notes]
      properties:
        contract_id:
          type: string
        notes:
          type: array
          items:
            $ref: "#/components/schemas/ContractNote"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: object
              required: [code, message]
              properties:
                code:
                  type: string
                  enum: [NOT_FOUND, INVALID_INPUT, INTERNAL, RATE_LIMITED, UNSUPPORTED_MEDIA_TYPE]
                message:
                  type: string
                request_id:
                  type: string
            - type: object
              properties:
                error:
                  type: string
...
    RoleError:
      type: object
      required: [error, required, role]
      properties:
        error:
          type: string
        required:
          type: string
        role:
          type: string
    Contract:
      type: object
      required: [id, network, label, wasm_hash, created_at_ledger, backfill_complete_at, status, added_at, tags]
      properties:
        id:
          type: string
        network:
          type: string
        label:
          type: string
        wasm_hash:
          type: string
        created_at_ledger:
          type: integer
          format: int64
        backfill_complete_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, active, backfilling, error, paused]
        added_at:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
          description: User-defined tags, sorted ascending. Always present; empty when the contract has no tags.
...